+++ cli/src/commands/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2019-2022 Parity Technologies (UK) Ltd.
+// This file is dual-licensed as Apache-2.0 or GPL-3.0.
+// see LICENSE for license details.
+
+pub mod codegen;
+pub mod compatibility;
+pub mod metadata;
